+++ docs/changelog.rst
@@ -2,6 +2,11 @@
 Changelog
 =========
 
+* :bug:`5124` Users will now correctly see all the events related to lending in the defi view.
+* :bug:`5126` APR and APY for borrowing and lending in Aave should properly show again.
+* :bug:`5128` Ethereum transactions where no value was transferred will now be correctly decoded.
+* :bug:`-` Uniswap V3 oracle will now correctly skip assets with no liquidity when querying prices.
+
 * :release:`1.26.1 <2022-11-04>`
 * :feature:`5114` Add an option to reset the global DB assets that are known by the app to their original state. A so called soft-reset.
 * :feature:`5102` Show stability pool to liquity module.